feat(bag): move BagDatabase from deriva.core to deriva.bag
Relocate the schema-independent BDBag-to-SQLAlchemy reader from
deriva.core.bag_database to its canonical home in
deriva.bag.database, alongside the rest of the bag-oriented
machinery added in the previous commit.

Changes:
- deriva/bag/database.py: new canonical location. Switched the
  engine factory from bare create_engine() to
  deriva.bag.sqlite_helpers.create_wal_engine, which picks up
  WAL journal mode, synchronous=NORMAL, foreign_keys=ON, and
  busy_timeout=5000 from one shared place. Added a call to
  ensure_schema_meta(engine, BAG_SCHEMA_VERSION) so a cache
  written by a future deriva.bag version raises SchemaVersionError
  on open rather than failing later with a confusing
  column-missing error.

- deriva/core/bag_database.py: replaced the 749-line
  implementation with a thin re-export shim that imports from
  deriva.bag.database. Existing callers (deriva-ml's DatabaseModel,
  user scripts) continue to work without changes. The shim is
  scheduled for removal after the deriva-ml migration PR lands
  and the in-tree consumers update their imports.

- deriva/core/__init__.py: dropped the top-level
  ``from deriva.core.bag_database import BagDatabase`` re-export.
  That line caused a circular import once BagDatabase moved to
  deriva.bag.database (which itself imports from
  deriva.core.ermrest_model, triggering deriva.core.__init__
  mid-import). Callers that did ``from deriva.core import
  BagDatabase`` should switch to ``from deriva.bag import
  BagDatabase`` or import via the legacy shim at
  ``deriva.core.bag_database`` (which still works).

- tests/deriva/bag/test_database.py: 12 tests covering schema
  load, row ingest with type coercion, ORM-class query, WAL mode
  pragma assertion (regression guard for the engine-factory
  swap), foreign_keys=ON, schema_meta record + forward-incompat
  raise, and asset URL localization (xfail: documents a
  pre-existing upstream bug in _localize_asset_row where the
  lookup compares full URLs against a path-keyed fetch-map; fix
  out of scope for this commit).

All 41 deriva.bag tests pass (1 xfail by design). The
non-network deriva.core test suite still passes (94 passed,
105 skipped — skips require a live catalog).
